requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
Flask-Bcrypt
Flask-WTF
Flask-JWT-Extended
Flask-Limiter
Flask-Mail
Flask-Cors
authlib
redis
boto3
itsdangerous
eventlet
Flask-SocketIO
gunicorn
python-dotenv
Pillow
moviepy
hachoir
celery
requests
elasticsearch
razorpay
